feat: project cleanup
