+++ TrollsVsElves/TrollsVsElves/Core/Components/Component.cs
@@ -1,0 +1,10 @@
+﻿using TrollsVsElves.Core.GameObjects;
+using TrollsVsElves.Core.Textures;
+
+namespace TrollsVsElves.Core.Components;
+
+public class Component
+{
+    public GameObject GameObject { get; set; }
+    public Transform Transform => GameObject.Transform;
+}
