+++ TrollsVsElves/TrollsVsElves/Core/Components/IDrawableComponent.cs
@@ -1,0 +1,6 @@
+﻿namespace TrollsVsElves.Core.Components;
+
+public interface IDrawableComponent
+{
+    void OnDraw();
+}
