+++ TrollsVsElves/TrollsVsElves/Core/Components/IUpdateableComponent.cs
@@ -1,0 +1,6 @@
+﻿namespace TrollsVsElves.Core.Components;
+
+public interface IUpdateableComponent
+{
+    void OnUpdate();
+}
