+++ TrollsVsElves/TrollsVsElves/Core/Lifetime/ISingleton.cs
@@ -1,0 +1,4 @@
+﻿namespace TrollsVsElves.Core.Lifetime;
+
+public interface ISingleton
+{ }
