+++ TrollsVsElves/TrollsVsElves/Core/Lifetime/ITransient.cs
@@ -1,0 +1,4 @@
+﻿namespace TrollsVsElves.Core.Lifetime;
+
+public interface ITransient
+{ }
