+++ TrollsVsElves/TrollsVsElves/Core/Sprites/SpriteRenderer.cs
@@ -1,9 +1,11 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using TrollsVsElves.Core.Components;
+using TrollsVsElves.Core.Lifetime;
 
-namespace TrollsVsElves
+namespace TrollsVsElves.Core.Sprites
 {
-    public class SpriteRenderer : Component, IDrawable, ITransient
+    public class SpriteRenderer : Component, Components.IDrawableComponent, ITransient
     {
         private Sprite _sprite;
         private SpriteBatch _spriteBatch;
