--- TrollsVsElves/TrollsVsElves/Scripts/Components/Component.cs
@@ -1,8 +1,0 @@
-﻿namespace TrollsVsElves
-{
-    public class Component
-    {
-        public Transform Transform => GameObject.Transform;
-        public GameObject GameObject { get; set; }
-    }
-}
