--- TrollsVsElves/TrollsVsElves/Scripts/Components/IDrawable.cs
@@ -1,7 +1,0 @@
-﻿namespace TrollsVsElves
-{
-    public interface IDrawable
-    {
-        void OnDraw();
-    }
-}
