--- TrollsVsElves/TrollsVsElves/Scripts/Components/IUpdateable.cs
@@ -1,7 +1,0 @@
-﻿namespace TrollsVsElves
-{
-    public interface IUpdateable
-    {
-        void OnUpdate();
-    }
-}
