+++ TrollsVsElves/TrollsVsElves/Scripts/Components/MoveWithCursorComponent.cs
@@ -1,0 +1,42 @@
+﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Input;
+using TrollsVsElves.Core.Components;
+using TrollsVsElves.Core.Lifetime;
+using TrollsVsElves.Core.Textures;
+using TrollsVsElves.Scripts.Services;
+
+namespace TrollsVsElves.Scripts.Components;
+
+public class MoveWithCursorComponent : Component, Core.Components.IUpdateableComponent, ITransient
+{
+    private InputHandlerService _inputHandler;
+    private int _movementSpeed;
+
+    public MoveWithCursorComponent(InputHandlerService inputHandler)
+    {
+        _movementSpeed = 400;
+        _inputHandler = inputHandler;
+    }
+
+    public void OnUpdate()
+    {
+        var mouseState = _inputHandler.MouseState;
+        var mousePosition = new Vector2(mouseState.X, mouseState.Y);
+
+        var position = Transform.Position;
+        var direciton = mousePosition - position;
+
+        var distance = Vector2.Distance(mousePosition, position);
+        direciton.Normalize();
+
+        if (distance < 5)
+        {
+            return;
+        }
+
+        if (mouseState.RightButton == ButtonState.Pressed)
+        {
+            Transform.Translate(direciton * _movementSpeed * Time.DeltaTime);
+        }
+    }
+}
