--- TrollsVsElves/TrollsVsElves/Scripts/InputHandler.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.Xna.Framework.Input;
-
-namespace TrollsVsElves
-{
-    public class InputHandler
-    {
-        public MouseState MouseState { get; set; }
-        public KeyboardState KeyboardState { get; set; }
-    }
-}
