--- TrollsVsElves/TrollsVsElves/Scripts/Services/Lifetime/ITransient.cs
@@ -1,4 +1,0 @@
-﻿namespace TrollsVsElves
-{
-    public interface ITransient { }
-}
