+++ TrollsVsElves/TrollsVsElves/Scripts/Time.cs
@@ -1,7 +1,6 @@
-﻿namespace TrollsVsElves
+﻿namespace TrollsVsElves.Scripts;
+
+public static class Time
 {
-    public static class Time
-    {
-        public static float DeltaTime { get; set; }
-    }
+    public static float DeltaTime { get; set; }
 }
